♻️  Refactor avatar serializers
