+++ src/sentry/api/bases/avatar.py
@@ -3,7 +3,6 @@ from rest_framework.response import Response
 
 from sentry.api.fields import AvatarField
 from sentry.api.serializers import serialize
-from sentry.utils.avatar import is_black_alpha_only
 
 
 class AvatarSerializer(serializers.Serializer):
@@ -26,34 +25,10 @@ class AvatarSerializer(serializers.Serializer):
         return attrs
 
 
-class SentryAppLogoSerializer(serializers.Serializer):
-    avatar_photo = AvatarField(required=False, is_sentry_app=True)
-    avatar_type = serializers.ChoiceField(choices=(("default", "default"), ("upload", "upload")))
-    color = serializers.BooleanField(required=True)
-
-    def validate(self, attrs):
-        attrs = super().validate(attrs)
-
-        if attrs.get("avatar_type") == "upload" and not attrs.get("avatar_photo"):
-            raise serializers.ValidationError({"avatar_photo": "A logo is required."})
-
-        if (
-            not attrs.get("color")
-            and attrs.get("avatar_type") == "upload"
-            and not is_black_alpha_only(attrs.get("avatar_photo"))
-        ):
-            raise serializers.ValidationError(
-                {
-                    "avatar_photo": "The icon must only use black and should contain an alpha channel."
-                }
-            )
-
-        return attrs
-
-
 class AvatarMixin:
     object_type = None
     model = None
+    serializer_cls = AvatarSerializer
 
     def get(self, request, **kwargs):
         obj = kwargs.pop(self.object_type, None)
@@ -68,11 +43,9 @@ class AvatarMixin:
     def put(self, request, **kwargs):
         obj = kwargs.pop(self.object_type, None)
 
-        SerializerCls = AvatarSerializer
-        if self.object_type == "sentry_app":
-            SerializerCls = SentryAppLogoSerializer
-
-        serializer = SerializerCls(data=request.data, context=self.get_serializer_context(obj))
+        serializer = self.serializer_cls(
+            data=request.data, context=self.get_serializer_context(obj)
+        )
 
         if not serializer.is_valid():
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
