+++ src/sentry/api/endpoints/sentry_app_avatar.py
@@ -1,11 +1,13 @@
 from sentry.api.bases import SentryAppBaseEndpoint
 from sentry.api.bases.avatar import AvatarMixin
+from sentry.api.serializers.rest_framework.sentry_app import SentryAppAvatarSerializer
 from sentry.models import SentryAppAvatar
 
 
 class SentryAppAvatarEndpoint(AvatarMixin, SentryAppBaseEndpoint):
     object_type = "sentry_app"
     model = SentryAppAvatar
+    serializer_cls = SentryAppAvatarSerializer
 
     def get(self, request, **kwargs):
         return super().get(request, access=request.access, **kwargs)
